# MultiCFSnapshot for NewIterators() API