--- a/db/db_impl/db_impl.h
+++ b/db/db_impl/db_impl.h
@@ -2355,18 +2355,20 @@ class DBImpl : public DB {
 
   // A structure to contain ColumnFamilyData and the SuperVersion obtained for
   // the consistent view of DB
-  struct ColumnFamilyDataSuperVersionPair {
+  struct ColumnFamilySuperVersionPair {
+    ColumnFamilyHandleImpl* cfh;
     ColumnFamilyData* cfd;
 
     // SuperVersion for the column family obtained in a manner that ensures a
     // consistent view across all column families in the DB
     SuperVersion* super_version;
-    ColumnFamilyDataSuperVersionPair(ColumnFamilyHandle* column_family,
-                                     SuperVersion* sv)
-        : cfd(static_cast<ColumnFamilyHandleImpl*>(column_family)->cfd()),
+    ColumnFamilySuperVersionPair(ColumnFamilyHandle* column_family,
+                                 SuperVersion* sv)
+        : cfh(static_cast<ColumnFamilyHandleImpl*>(column_family)),
+          cfd(cfh->cfd()),
           super_version(sv) {}
 
-    ColumnFamilyDataSuperVersionPair() = default;
+    ColumnFamilySuperVersionPair() = default;
   };
 
   // A common function to obtain a consistent snapshot, which can be implicit
@@ -2383,6 +2385,13 @@ class DBImpl : public DB {
   // `sv_from_thread_local` being set to false indicates that the SuperVersion
   // obtained from the ColumnFamilyData, whereas true indicates they are thread
   // local.
+  // `sv_exclusive_access` is used to indicate whether thread-local SuperVersion
+  // can be obtained without extra ref (by GetAndRefSuperVersion()) or not
+  // (by GetReferencedSuperVersion()). For instance, point lookup like MultiGet
+  // does not require SuperVersion to be re-acquired throughout the entire
+  // invocation (no need extra ref), while MultiCfIterators may need the
+  // SuperVersion to be updated during Refresh() (requires extra ref).
+  //
   // A non-OK status will be returned if for a column family that enables
   // user-defined timestamp feature, the specified `ReadOptions.timestamp`
   // attemps to read collapsed history.
@@ -2390,7 +2399,8 @@ class DBImpl : public DB {
   Status MultiCFSnapshot(const ReadOptions& read_options,
                          ReadCallback* callback,
                          IterDerefFuncType iter_deref_func, T* cf_list,
-                         SequenceNumber* snapshot, bool* sv_from_thread_local);
+                         SequenceNumber* snapshot, bool* sv_from_thread_local,
+                         bool sv_exclusive_access);
 
   // The actual implementation of the batching MultiGet. The caller is expected
   // to have acquired the SuperVersion and pass in a snapshot sequence number