--- a/unreleased_history/bug_fixes/consistent_view_for_new_iterators_api.md
+++ b/unreleased_history/bug_fixes/consistent_view_for_new_iterators_api.md
@@ -1,0 +1,1 @@
+Provide consistent view of the database across the column families for `NewIterators()` API.